fix: Prevent incorrect draft loading with precise recipient matching
