+++ frontend/app/src/main/java/com/fiveis/xend/ui/inbox/InboxActivity.kt
@@ -24,6 +24,7 @@ import com.fiveis.xend.ui.search.SearchActivity
 import com.fiveis.xend.ui.sent.SentActivity
 import com.fiveis.xend.ui.theme.XendTheme
 import com.fiveis.xend.ui.view.MailDetailActivity
+import com.fiveis.xend.utils.EmailUtils
 
 class InboxActivity : ComponentActivity() {
     private val viewModel: InboxViewModel by viewModels { InboxViewModelFactory(this.applicationContext) }
@@ -85,8 +86,8 @@ class InboxActivity : ComponentActivity() {
                 if (uiState.showAddContactDialog) {
                     uiState.selectedEmailForContact?.let { email ->
                         AddContactDialog(
-                            senderName = extractSenderName(email.fromEmail),
-                            senderEmail = extractSenderEmail(email.fromEmail),
+                            senderName = EmailUtils.extractSenderName(email.fromEmail),
+                            senderEmail = EmailUtils.extractEmailAddress(email.fromEmail),
                             groups = uiState.groups,
                             onDismiss = { viewModel.dismissAddContactDialog() },
                             onConfirm = { name, emailAddr, senderRole, recipientRole, personalPrompt, groupId ->
@@ -104,21 +105,6 @@ class InboxActivity : ComponentActivity() {
                 }
             }
         }
-    }
-
-    private fun extractSenderName(fromEmail: String): String {
-        val nameRegex = "(.+?)\\s*<".toRegex()
-        val matchResult = nameRegex.find(fromEmail)
-        val name = matchResult?.groupValues?.get(1)?.trim()
-            ?: fromEmail.substringBefore("<").trim().ifEmpty { fromEmail }
-        // Remove surrounding quotes if present
-        return name.trim('"', '\'')
-    }
-
-    private fun extractSenderEmail(fromEmail: String): String {
-        val emailRegex = "<(.+?)>".toRegex()
-        val matchResult = emailRegex.find(fromEmail)
-        return matchResult?.groupValues?.get(1)?.trim() ?: fromEmail
     }
 }
